+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- hnefatafl-client: handle whitespace in a password better.
+
 ## [1.1.4] - 2025-05-23
 
 - hnefatafl-client: don't limit the valid password characters.